true file level opt-in
Previously we checked the comments leading to any top level symbol
for the opt in pragma this caused the file to be skipped if any of
the top level symbols was missing the opt in pragma. Now, if any of
the top level comments has the opt in pragma and the environments
match the file will be included.